+++ DifferentiationInterfaceTest/src/tests/correctness_eval.jl
@@ -1,3 +1,6 @@
+has_size(::Union{Number,AbstractArray}) = true
+has_size(_x) = false
+
 const PME = PreparationMismatchError
 
 for op in ALL_OPS
@@ -49,32 +52,19 @@ for op in ALL_OPS
             rtol::Real,
             scenario_intact::Bool,
             sparsity::Bool,
+            reprepare::Bool,
         )
-            (; f, x, y, res1, contexts, smaller) = new_scen = deepcopy(scen)
-            xrand = myrandom(x)
-            rewrap = Rewrap(contexts...)
-            contextsrand = rewrap(map(myrandom ∘ unwrap, contexts)...)
+            (; f, x, y, res1, contexts, prep_args) = new_scen = deepcopy(scen)
             local prepstrict
             preptup_cands_val, preptup_cands_noval = map(1:2) do _
-                new_smaller = if isnothing(smaller) || adapt_batchsize(ba, smaller) != ba
-                        deepcopy(scen)
-                    else
-                        deepcopy(smaller)
-                    end
-                prep = $prep_op(f, ba, xrand, contextsrand...)
-                prepstrict = $prep_op!(
-                    f,
-                    $prep_op(
-                        new_smaller.f,
-                        ba,
-                        new_smaller.x,
-                        new_smaller.contexts...;
-                        strict=Val(true),
-                    ),
-                    ba,
-                    xrand,
-                    contextsrand...,
+                prep = $prep_op(f, ba, prep_args.x, prep_args.contexts...)
+                prepstrict = $prep_op(
+                    f, ba, prep_args.x, prep_args.contexts...; strict=Val(true)
                 )
+                if reprepare && has_size(x) && has_size(y) && (size(x) != size(prep_args.x))
+                    prep = $prep_op!(f, prep, ba, x, contexts...)
+                    prepstrict = $prep_op!(f, prepstrict, ba, x, contexts...)
+                end
                 [(), (prep,), (prepstrict,)]
             end
             for (preptup_val, preptup_noval) in zip(preptup_cands_val, preptup_cands_noval)
@@ -116,32 +106,19 @@ for op in ALL_OPS
             rtol::Real,
             scenario_intact::Bool,
             sparsity::Bool,
+            reprepare::Bool,
         )
-            (; f, x, y, res1, contexts, smaller) = new_scen = deepcopy(scen)
-            xrand = myrandom(x)
-            rewrap = Rewrap(contexts...)
-            contextsrand = rewrap(map(myrandom ∘ unwrap, contexts)...)
+            (; f, x, y, res1, contexts, prep_args) = new_scen = deepcopy(scen)
             local prepstrict
             preptup_cands_val, preptup_cands_noval = map(1:2) do _
-                new_smaller = if isnothing(smaller) || adapt_batchsize(ba, smaller) != ba
-                        deepcopy(scen)
-                    else
-                        deepcopy(smaller)
-                    end
-                prep = $prep_op(f, ba, xrand, contextsrand...)
-                prepstrict = $prep_op!(
-                    f,
-                    $prep_op(
-                        new_smaller.f,
-                        ba,
-                        new_smaller.x,
-                        new_smaller.contexts...;
-                        strict=Val(true),
-                    ),
-                    ba,
-                    xrand,
-                    contextsrand...,
+                prep = $prep_op(f, ba, prep_args.x, prep_args.contexts...)
+                prepstrict = $prep_op(
+                    f, ba, prep_args.x, prep_args.contexts...; strict=Val(true)
                 )
+                if reprepare && has_size(x) && has_size(y) && (size(x) != size(prep_args.x))
+                    prep = $prep_op!(f, prep, ba, x, contexts...)
+                    prepstrict = $prep_op!(f, prepstrict, ba, x, contexts...)
+                end
                 [(), (prep,), (prepstrict,)]
             end
             for (preptup_val, preptup_noval) in zip(preptup_cands_val, preptup_cands_noval)
@@ -199,34 +176,27 @@ for op in ALL_OPS
             rtol::Real,
             scenario_intact::Bool,
             sparsity::Bool,
+            reprepare::Bool,
         )
-            (; f, x, y, res1, contexts, smaller) = new_scen = deepcopy(scen)
-            xrand, yrand = myrandom(x), myrandom(y)
-            rewrap = Rewrap(contexts...)
-            contextsrand = rewrap(map(myrandom ∘ unwrap, contexts)...)
+            (; f, x, y, res1, contexts, prep_args) = new_scen = deepcopy(scen)
             local prepstrict
             preptup_cands_val, preptup_cands_noval = map(1:2) do _
-                new_smaller = if isnothing(smaller) || adapt_batchsize(ba, smaller) != ba
-                        deepcopy(scen)
-                    else
-                        deepcopy(smaller)
-                    end
-                prep = $prep_op(f, copy(yrand), ba, xrand, contextsrand...)
-                prepstrict = $prep_op!(
+                prep = $prep_op(f, prep_args.y, ba, prep_args.x, prep_args.contexts...)
+                prepstrict = $prep_op(
                     f,
-                    copy(yrand),
-                    $prep_op(
-                        new_smaller.f,
-                        copy(new_smaller.y),
-                        ba,
-                        new_smaller.x,
-                        new_smaller.contexts...;
-                        strict=Val(true),
-                    ),
+                    prep_args.y,
                     ba,
-                    xrand,
-                    contextsrand...,
+                    prep_args.x,
+                    prep_args.contexts...;
+                    strict=Val(true),
                 )
+                if reprepare &&
+                    has_size(x) &&
+                    has_size(y) &&
+                    (size(x) != size(prep_args.x) || size(y) != prep_args.y)
+                    prep = $prep_op!(f, y, prep, ba, x, contexts...)
+                    prepstrict = $prep_op!(f, y, prepstrict, ba, x, contexts...)
+                end
                 [(), (prep,), (prepstrict,)]
             end
             for (preptup_val, preptup_noval) in zip(preptup_cands_val, preptup_cands_noval)
@@ -276,34 +246,27 @@ for op in ALL_OPS
             rtol::Real,
             scenario_intact::Bool,
             sparsity::Bool,
+            reprepare::Bool,
         )
-            (; f, x, y, res1, contexts, smaller) = new_scen = deepcopy(scen)
-            xrand, yrand = myrandom(x), myrandom(y)
-            rewrap = Rewrap(contexts...)
-            contextsrand = rewrap(map(myrandom ∘ unwrap, contexts)...)
+            (; f, x, y, res1, contexts, prep_args) = new_scen = deepcopy(scen)
             local prepstrict
             preptup_cands_val, preptup_cands_noval = map(1:2) do _
-                new_smaller = if isnothing(smaller) || adapt_batchsize(ba, smaller) != ba
-                        deepcopy(scen)
-                    else
-                        deepcopy(smaller)
-                    end
-                prep = $prep_op(f, copy(yrand), ba, xrand, contextsrand...)
-                prepstrict = $prep_op!(
+                prep = $prep_op(f, prep_args.y, ba, prep_args.x, prep_args.contexts...)
+                prepstrict = $prep_op(
                     f,
-                    copy(yrand),
-                    $prep_op(
-                        new_smaller.f,
-                        copy(new_smaller.y),
-                        ba,
-                        new_smaller.x,
-                        new_smaller.contexts...;
-                        strict=Val(true),
-                    ),
+                    prep_args.y,
                     ba,
-                    xrand,
-                    contextsrand...,
+                    prep_args.x,
+                    prep_args.contexts...;
+                    strict=Val(true),
                 )
+                if reprepare &&
+                    has_size(x) &&
+                    has_size(y) &&
+                    (size(x) != size(prep_args.x) || size(y) != prep_args.y)
+                    prep = $prep_op!(f, y, prep, ba, x, contexts...)
+                    prepstrict = $prep_op!(f, y, prepstrict, ba, x, contexts...)
+                end
                 [(), (prep,), (prepstrict,)]
             end
             for (preptup_val, preptup_noval) in zip(preptup_cands_val, preptup_cands_noval)
@@ -364,32 +327,19 @@ for op in ALL_OPS
             rtol::Real,
             scenario_intact::Bool,
             sparsity::Bool,
+            reprepare::Bool,
         )
-            (; f, x, y, res1, res2, contexts, smaller) = new_scen = deepcopy(scen)
-            xrand = myrandom(x)
-            rewrap = Rewrap(contexts...)
-            contextsrand = rewrap(map(myrandom ∘ unwrap, contexts)...)
+            (; f, x, y, res1, res2, contexts, prep_args) = new_scen = deepcopy(scen)
             local prepstrict
             preptup_cands_val, preptup_cands_noval = map(1:2) do _
-                new_smaller = if isnothing(smaller) || adapt_batchsize(ba, smaller) != ba
-                        deepcopy(scen)
-                    else
-                        deepcopy(smaller)
-                    end
-                prep = $prep_op(f, ba, xrand, contextsrand...)
-                prepstrict = $prep_op!(
-                    f,
-                    $prep_op(
-                        new_smaller.f,
-                        ba,
-                        new_smaller.x,
-                        new_smaller.contexts...;
-                        strict=Val(true),
-                    ),
-                    ba,
-                    xrand,
-                    contextsrand...,
+                prep = $prep_op(f, ba, prep_args.x, prep_args.contexts...)
+                prepstrict = $prep_op(
+                    f, ba, prep_args.x, prep_args.contexts...; strict=Val(true)
                 )
+                if reprepare && has_size(x) && has_size(y) && (size(x) != size(prep_args.x))
+                    prep = $prep_op!(f, prep, ba, x, contexts...)
+                    prepstrict = $prep_op!(f, prepstrict, ba, x, contexts...)
+                end
                 [(), (prep,), (prepstrict,)]
             end
             for (preptup_val, preptup_noval) in zip(preptup_cands_val, preptup_cands_noval)
@@ -433,32 +383,19 @@ for op in ALL_OPS
             rtol::Real,
             scenario_intact::Bool,
             sparsity::Bool,
+            reprepare::Bool,
         )
-            (; f, x, y, res1, res2, contexts, smaller) = new_scen = deepcopy(scen)
-            xrand = myrandom(x)
-            rewrap = Rewrap(contexts...)
-            contextsrand = rewrap(map(myrandom ∘ unwrap, contexts)...)
+            (; f, x, y, res1, res2, contexts, prep_args) = new_scen = deepcopy(scen)
             local prepstrict
             preptup_cands_val, preptup_cands_noval = map(1:2) do _
-                new_smaller = if isnothing(smaller) || adapt_batchsize(ba, smaller) != ba
-                        deepcopy(scen)
-                    else
-                        deepcopy(smaller)
-                    end
-                prep = $prep_op(f, ba, xrand, contextsrand...)
-                prepstrict = $prep_op!(
-                    f,
-                    $prep_op(
-                        new_smaller.f,
-                        ba,
-                        new_smaller.x,
-                        new_smaller.contexts...;
-                        strict=Val(true),
-                    ),
-                    ba,
-                    xrand,
-                    contextsrand...,
+                prep = $prep_op(f, ba, prep_args.x, prep_args.contexts...)
+                prepstrict = $prep_op(
+                    f, ba, prep_args.x, prep_args.contexts...; strict=Val(true)
                 )
+                if reprepare && has_size(x) && has_size(y) && (size(x) != size(prep_args.x))
+                    prep = $prep_op!(f, prep, ba, x, contexts...)
+                    prepstrict = $prep_op!(f, prepstrict, ba, x, contexts...)
+                end
                 [(), (prep,), (prepstrict,)]
             end
             for (preptup_val, preptup_noval) in zip(preptup_cands_val, preptup_cands_noval)
@@ -519,46 +456,36 @@ for op in ALL_OPS
             rtol::Real,
             scenario_intact::Bool,
             sparsity::Bool,
+            reprepare::Bool,
         )
-            (; f, x, y, tang, res1, contexts, smaller) = new_scen = deepcopy(scen)
-            xrand, tangrand = myrandom(x), myrandom(tang)
-            rewrap = Rewrap(contexts...)
-            contextsrand = rewrap(map(myrandom ∘ unwrap, contexts)...)
+            (; f, x, y, t, res1, contexts, prep_args) = new_scen = deepcopy(scen)
             local prepstrict
             preptup_cands_val, preptup_cands_noval = map(1:2) do _
-                new_smaller = if isnothing(smaller) || adapt_batchsize(ba, smaller) != ba
-                        deepcopy(scen)
-                    else
-                        deepcopy(smaller)
-                    end
-                prep = $prep_op(f, ba, xrand, tangrand, contextsrand...)
-                prepstrict = $prep_op!(
+                prep = $prep_op(f, ba, prep_args.x, prep_args.t, prep_args.contexts...)
+                prepstrict = $prep_op(
                     f,
-                    $prep_op(
-                        new_smaller.f,
-                        ba,
-                        new_smaller.x,
-                        new_smaller.tang,
-                        new_smaller.contexts...;
-                        strict=Val(true),
-                    ),
                     ba,
-                    xrand,
-                    tangrand,
-                    contextsrand...,
+                    prep_args.x,
+                    prep_args.t,
+                    prep_args.contexts...;
+                    strict=Val(true),
                 )
-                prep_same = $prep_op_same(f, ba, x, tangrand, contexts...)
+                prep_same = $prep_op_same(f, ba, x, map(zero, t), contexts...)
+                if reprepare && has_size(x) && has_size(y) && (size(x) != size(prep_args.x))
+                    prep = $prep_op!(f, prep, ba, x, t, contexts...)
+                    prepstrict = $prep_op!(f, prepstrict, ba, x, t, contexts...)
+                end
                 [(), (prep,), (prepstrict,), (prep_same,)]
             end
             for (preptup_val, preptup_noval) in zip(preptup_cands_val, preptup_cands_noval)
                 y_out1_val, res1_out1_val = $val_and_op(
-                    f, preptup_val..., ba, x, tang, contexts...
+                    f, preptup_val..., ba, x, t, contexts...
                 )
                 y_out2_val, res1_out2_val = $val_and_op(
-                    f, preptup_val..., ba, x, tang, contexts...
+                    f, preptup_val..., ba, x, t, contexts...
                 )
-                res1_out1_noval = $op(f, preptup_noval..., ba, x, tang, contexts...)
-                res1_out2_noval = $op(f, preptup_noval..., ba, x, tang, contexts...)
+                res1_out1_noval = $op(f, preptup_noval..., ba, x, t, contexts...)
+                res1_out2_noval = $op(f, preptup_noval..., ba, x, t, contexts...)
                 let (≈)(x, y) = isapprox(x, y; atol, rtol)
                     @test isempty(preptup_noval) || only(preptup_noval) isa $P
                     @test y_out1_val ≈ scen.y
@@ -571,8 +498,8 @@ for op in ALL_OPS
                     end
                 end
             end
-            @test_throws PME $val_and_op(nothing, prepstrict, ba, x, tang, contexts...)
-            @test_throws PME $op(nothing, prepstrict, ba, x, tang, contexts...)
+            @test_throws PME $val_and_op(nothing, prepstrict, ba, x, t, contexts...)
+            @test_throws PME $op(nothing, prepstrict, ba, x, t, contexts...)
             scenario_intact && @test new_scen == scen
             return nothing
         end
@@ -585,35 +512,25 @@ for op in ALL_OPS
             rtol::Real,
             scenario_intact::Bool,
             sparsity::Bool,
+            reprepare::Bool,
         )
-            (; f, x, y, tang, res1, contexts, smaller) = new_scen = deepcopy(scen)
-            xrand, tangrand = myrandom(x), myrandom(tang)
-            rewrap = Rewrap(contexts...)
-            contextsrand = rewrap(map(myrandom ∘ unwrap, contexts)...)
+            (; f, x, y, t, res1, contexts, prep_args) = new_scen = deepcopy(scen)
             local prepstrict
             preptup_cands_val, preptup_cands_noval = map(1:2) do _
-                new_smaller = if isnothing(smaller) || adapt_batchsize(ba, smaller) != ba
-                        deepcopy(scen)
-                    else
-                        deepcopy(smaller)
-                    end
-                prep = $prep_op(f, ba, xrand, tangrand, contextsrand...)
-                prepstrict = $prep_op!(
+                prep = $prep_op(f, ba, prep_args.x, prep_args.t, prep_args.contexts...)
+                prepstrict = $prep_op(
                     f,
-                    $prep_op(
-                        new_smaller.f,
-                        ba,
-                        new_smaller.x,
-                        new_smaller.tang,
-                        new_smaller.contexts...;
-                        strict=Val(true),
-                    ),
                     ba,
-                    xrand,
-                    tangrand,
-                    contextsrand...,
+                    prep_args.x,
+                    prep_args.t,
+                    prep_args.contexts...;
+                    strict=Val(true),
                 )
-                prep_same = $prep_op_same(f, ba, x, tangrand, contexts...)
+                prep_same = $prep_op_same(f, ba, x, map(zero, t), contexts...)
+                if reprepare && has_size(x) && has_size(y) && (size(x) != size(prep_args.x))
+                    prep = $prep_op!(f, prep, ba, x, t, contexts...)
+                    prepstrict = $prep_op!(f, prepstrict, ba, x, t, contexts...)
+                end
                 [(), (prep,), (prepstrict,), (prep_same,)]
             end
             for (preptup_val, preptup_noval) in zip(preptup_cands_val, preptup_cands_noval)
@@ -622,38 +539,38 @@ for op in ALL_OPS
                 res1_in1_noval = mysimilar(res1)
                 res1_in2_noval = mysimilar(res1)
                 y_out1_val, res1_out1_val = $val_and_op!(
-                    f, res1_in1_val, preptup_val..., ba, x, tang, contexts...
+                    f, res1_in1_val, preptup_val..., ba, x, t, contexts...
                 )
                 y_out2_val, res1_out2_val = $val_and_op!(
-                    f, res1_in2_val, preptup_val..., ba, x, tang, contexts...
+                    f, res1_in2_val, preptup_val..., ba, x, t, contexts...
                 )
                 res1_out1_noval = $op!(
-                    f, res1_in1_noval, preptup_noval..., ba, x, tang, contexts...
+                    f, res1_in1_noval, preptup_noval..., ba, x, t, contexts...
                 )
                 res1_out2_noval = $op!(
-                    f, res1_in2_noval, preptup_noval..., ba, x, tang, contexts...
+                    f, res1_in2_noval, preptup_noval..., ba, x, t, contexts...
                 )
                 let (≈)(x, y) = isapprox(x, y; atol, rtol)
                     @test isempty(preptup_noval) || only(preptup_noval) isa $P
                     @test y_out1_val ≈ scen.y
                     @test y_out2_val ≈ scen.y
+                    @test res1_in1_val === res1_out1_val
+                    @test res1_in2_val === res1_out2_val
+                    @test res1_in1_noval === res1_out1_noval
+                    @test res1_in2_noval === res1_out2_noval
                     for b in eachindex(scen.res1)
-                        @test res1_in1_val[b] === res1_out1_val[b]
-                        @test res1_in2_val[b] === res1_out2_val[b]
                         @test res1_out1_val[b] ≈ scen.res1[b]
                         @test res1_out2_val[b] ≈ scen.res1[b]
-                        @test res1_in1_noval[b] === res1_out1_noval[b]
-                        @test res1_in2_noval[b] === res1_out2_noval[b]
                         @test res1_out1_noval[b] ≈ scen.res1[b]
                         @test res1_out2_noval[b] ≈ scen.res1[b]
                     end
                 end
             end
             @test_throws PME $val_and_op!(
-                nothing, mysimilar(res1), prepstrict, ba, x, tang, contexts...
+                nothing, mysimilar(res1), prepstrict, ba, x, t, contexts...
             )
             @test_throws PME $op!(
-                nothing, mysimilar(res1), prepstrict, ba, x, tang, contexts...
+                nothing, mysimilar(res1), prepstrict, ba, x, t, contexts...
             )
             scenario_intact && @test new_scen == scen
             return nothing
@@ -667,37 +584,31 @@ for op in ALL_OPS
             rtol::Real,
             scenario_intact::Bool,
             sparsity::Bool,
+            reprepare::Bool,
         )
-            (; f, x, y, tang, res1, contexts, smaller) = new_scen = deepcopy(scen)
-            xrand, yrand, tangrand = myrandom(x), myrandom(y), myrandom(tang)
-            rewrap = Rewrap(contexts...)
-            contextsrand = rewrap(map(myrandom ∘ unwrap, contexts)...)
+            (; f, x, y, t, res1, contexts, prep_args) = new_scen = deepcopy(scen)
             local prepstrict
             preptup_cands_val, preptup_cands_noval = map(1:2) do _
-                new_smaller = if isnothing(smaller) || adapt_batchsize(ba, smaller) != ba
-                        deepcopy(scen)
-                    else
-                        deepcopy(smaller)
-                    end
-                prep = $prep_op(f, copy(yrand), ba, xrand, tangrand, contextsrand...)
-                prepstrict = $prep_op!(
-                    f,
-                    copy(yrand),
-                    $prep_op(
-                        new_smaller.f,
-                        copy(new_smaller.y),
-                        ba,
-                        new_smaller.x,
-                        new_smaller.tang,
-                        new_smaller.contexts...;
-                        strict=Val(true),
-                    ),
-                    ba,
-                    xrand,
-                    tangrand,
-                    contextsrand...,
+                prep = $prep_op(
+                    f, prep_args.y, ba, prep_args.x, prep_args.t, prep_args.contexts...
                 )
-                prep_same = $prep_op_same(f, copy(yrand), ba, x, tangrand, contexts...)
+                prepstrict = $prep_op(
+                    f,
+                    prep_args.y,
+                    ba,
+                    prep_args.x,
+                    prep_args.t,
+                    prep_args.contexts...;
+                    strict=Val(true),
+                )
+                prep_same = $prep_op_same(f, y, ba, x, map(zero, t), contexts...)
+                if reprepare &&
+                    has_size(x) &&
+                    has_size(y) &&
+                    (size(x) != size(prep_args.x) || size(y) != prep_args.y)
+                    prep = $prep_op!(f, y, prep, ba, x, t, contexts...)
+                    prepstrict = $prep_op!(f, y, prepstrict, ba, x, t, contexts...)
+                end
                 [(), (prep,), (prepstrict,), (prep_same,)]
             end
             for (preptup_val, preptup_noval) in zip(preptup_cands_val, preptup_cands_noval)
@@ -706,16 +617,16 @@ for op in ALL_OPS
                 y_in1_noval = mysimilar(y)
                 y_in2_noval = mysimilar(y)
                 y_out1_val, res1_out1_val = $val_and_op(
-                    f, y_in1_val, preptup_val..., ba, x, tang, contexts...
+                    f, y_in1_val, preptup_val..., ba, x, t, contexts...
                 )
                 y_out2_val, res1_out2_val = $val_and_op(
-                    f, y_in2_val, preptup_val..., ba, x, tang, contexts...
+                    f, y_in2_val, preptup_val..., ba, x, t, contexts...
                 )
                 res1_out1_noval = $op(
-                    f, y_in1_noval, preptup_noval..., ba, x, tang, contexts...
+                    f, y_in1_noval, preptup_noval..., ba, x, t, contexts...
                 )
                 res1_out2_noval = $op(
-                    f, y_in2_noval, preptup_noval..., ba, x, tang, contexts...
+                    f, y_in2_noval, preptup_noval..., ba, x, t, contexts...
                 )
                 let (≈)(x, y) = isapprox(x, y; atol, rtol)
                     @test isempty(preptup_noval) || only(preptup_noval) isa $P
@@ -732,11 +643,9 @@ for op in ALL_OPS
                 end
             end
             @test_throws PME $val_and_op(
-                nothing, mysimilar(y), prepstrict, ba, x, tang, contexts...
+                nothing, mysimilar(y), prepstrict, ba, x, t, contexts...
             )
-            @test_throws PME $op(
-                nothing, mysimilar(y), prepstrict, ba, x, tang, contexts...
-            )
+            @test_throws PME $op(nothing, mysimilar(y), prepstrict, ba, x, t, contexts...)
             scenario_intact && @test new_scen == scen
             return nothing
         end
@@ -749,37 +658,31 @@ for op in ALL_OPS
             rtol::Real,
             scenario_intact::Bool,
             sparsity::Bool,
+            reprepare::Bool,
         )
-            (; f, x, y, tang, res1, contexts, smaller) = new_scen = deepcopy(scen)
-            xrand, yrand, tangrand = myrandom(x), myrandom(y), myrandom(tang)
-            rewrap = Rewrap(contexts...)
-            contextsrand = rewrap(map(myrandom ∘ unwrap, contexts)...)
+            (; f, x, y, t, res1, contexts, prep_args) = new_scen = deepcopy(scen)
             local prepstrict
             preptup_cands_val, preptup_cands_noval = map(1:2) do _
-                new_smaller = if isnothing(smaller) || adapt_batchsize(ba, smaller) != ba
-                        deepcopy(scen)
-                    else
-                        deepcopy(smaller)
-                    end
-                prep = $prep_op(f, copy(yrand), ba, xrand, tangrand, contextsrand...)
-                prepstrict = $prep_op!(
-                    f,
-                    copy(yrand),
-                    $prep_op(
-                        new_smaller.f,
-                        copy(new_smaller.y),
-                        ba,
-                        new_smaller.x,
-                        new_smaller.tang,
-                        new_smaller.contexts...;
-                        strict=Val(true),
-                    ),
-                    ba,
-                    xrand,
-                    tangrand,
-                    contextsrand...,
+                prep = $prep_op(
+                    f, prep_args.y, ba, prep_args.x, prep_args.t, prep_args.contexts...
                 )
-                prep_same = $prep_op_same(f, copy(yrand), ba, x, tangrand, contexts...)
+                prepstrict = $prep_op(
+                    f,
+                    prep_args.y,
+                    ba,
+                    prep_args.x,
+                    prep_args.t,
+                    prep_args.contexts...;
+                    strict=Val(true),
+                )
+                prep_same = $prep_op_same(f, y, ba, x, map(zero, t), contexts...)
+                if reprepare &&
+                    has_size(x) &&
+                    has_size(y) &&
+                    (size(x) != size(prep_args.x) || size(y) != prep_args.y)
+                    prep = $prep_op!(f, y, prep, ba, x, t, contexts...)
+                    prepstrict = $prep_op!(f, y, prepstrict, ba, x, t, contexts...)
+                end
                 [(), (prep,), (prepstrict,), (prep_same,)]
             end
             for (preptup_val, preptup_noval) in zip(preptup_cands_val, preptup_cands_noval)
@@ -788,30 +691,16 @@ for op in ALL_OPS
                 y_in1_noval, res1_in1_noval = mysimilar(y), mysimilar(res1)
                 y_in2_noval, res1_in2_noval = mysimilar(y), mysimilar(res1)
                 y_out1_val, res1_out1_val = $val_and_op!(
-                    f, y_in1_val, res1_in1_val, preptup_val..., ba, x, tang, contexts...
+                    f, y_in1_val, res1_in1_val, preptup_val..., ba, x, t, contexts...
                 )
                 y_out2_val, res1_out2_val = $val_and_op!(
-                    f, y_in2_val, res1_in2_val, preptup_val..., ba, x, tang, contexts...
+                    f, y_in2_val, res1_in2_val, preptup_val..., ba, x, t, contexts...
                 )
                 res1_out1_noval = $op!(
-                    f,
-                    y_in1_noval,
-                    res1_in1_noval,
-                    preptup_noval...,
-                    ba,
-                    x,
-                    tang,
-                    contexts...,
+                    f, y_in1_noval, res1_in1_noval, preptup_noval..., ba, x, t, contexts...
                 )
                 res1_out2_noval = $op!(
-                    f,
-                    y_in2_noval,
-                    res1_in2_noval,
-                    preptup_noval...,
-                    ba,
-                    x,
-                    tang,
-                    contexts...,
+                    f, y_in2_noval, res1_in2_noval, preptup_noval..., ba, x, t, contexts...
                 )
                 let (≈)(x, y) = isapprox(x, y; atol, rtol)
                     @test isempty(preptup_noval) || only(preptup_noval) isa $P
@@ -819,23 +708,23 @@ for op in ALL_OPS
                     @test y_in2_val === y_out2_val
                     @test y_out1_val ≈ scen.y
                     @test y_out2_val ≈ scen.y
+                    @test res1_in1_val === res1_out1_val
+                    @test res1_in2_val === res1_out2_val
+                    @test res1_in1_noval === res1_out1_noval
+                    @test res1_in2_noval === res1_out2_noval
                     for b in eachindex(scen.res1)
-                        @test res1_in1_val[b] === res1_out1_val[b]
-                        @test res1_in2_val[b] === res1_out2_val[b]
                         @test res1_out1_val[b] ≈ scen.res1[b]
                         @test res1_out2_val[b] ≈ scen.res1[b]
-                        @test res1_in1_noval[b] === res1_out1_noval[b]
-                        @test res1_in2_noval[b] === res1_out2_noval[b]
                         @test res1_out1_noval[b] ≈ scen.res1[b]
                         @test res1_out2_noval[b] ≈ scen.res1[b]
                     end
                 end
             end
             @test_throws PME $val_and_op!(
-                nothing, mysimilar(y), mysimilar(res1), prepstrict, ba, x, tang, contexts...
+                nothing, mysimilar(y), mysimilar(res1), prepstrict, ba, x, t, contexts...
             )
             @test_throws PME $op!(
-                nothing, mysimilar(y), mysimilar(res1), prepstrict, ba, x, tang, contexts...
+                nothing, mysimilar(y), mysimilar(res1), prepstrict, ba, x, t, contexts...
             )
             scenario_intact && @test new_scen == scen
             return nothing
@@ -850,45 +739,35 @@ for op in ALL_OPS
             rtol::Real,
             scenario_intact::Bool,
             sparsity::Bool,
+            reprepare::Bool,
         )
-            (; f, x, y, tang, res1, res2, contexts, smaller) = new_scen = deepcopy(scen)
-            xrand, tangrand = myrandom(x), myrandom(tang)
-            rewrap = Rewrap(contexts...)
-            contextsrand = rewrap(map(myrandom ∘ unwrap, contexts)...)
+            (; f, x, y, t, res1, res2, contexts, prep_args) = new_scen = deepcopy(scen)
             local prepstrict
             preptup_cands_val, preptup_cands_noval = map(1:2) do _
-                new_smaller = if isnothing(smaller) || adapt_batchsize(ba, smaller) != ba
-                        deepcopy(scen)
-                    else
-                        deepcopy(smaller)
-                    end
-                prep = $prep_op(f, ba, xrand, tangrand, contextsrand...)
-                prepstrict = $prep_op!(
+                prep = $prep_op(f, ba, prep_args.x, prep_args.t, prep_args.contexts...)
+                prepstrict = $prep_op(
                     f,
-                    $prep_op(
-                        new_smaller.f,
-                        ba,
-                        new_smaller.x,
-                        new_smaller.tang,
-                        new_smaller.contexts...;
-                        strict=Val(true),
-                    ),
                     ba,
-                    xrand,
-                    tangrand,
-                    contextsrand...,
+                    prep_args.x,
+                    prep_args.t,
+                    prep_args.contexts...;
+                    strict=Val(true),
                 )
-                prep_same = $prep_op_same(f, ba, x, tangrand, contexts...)
+                prep_same = $prep_op_same(f, ba, x, map(zero, t), contexts...)
+                if reprepare && has_size(x) && has_size(y) && (size(x) != size(prep_args.x))
+                    prep = $prep_op!(f, prep, ba, x, t, contexts...)
+                    prepstrict = $prep_op!(f, prepstrict, ba, x, t, contexts...)
+                end
                 [(), (prep,), (prepstrict,), (prep_same,)]
             end
             for (preptup_val, preptup_noval) in zip(preptup_cands_val, preptup_cands_noval)
-                res2_out1_noval = $op(f, preptup_noval..., ba, x, tang, contexts...)
-                res2_out2_noval = $op(f, preptup_noval..., ba, x, tang, contexts...)
+                res2_out1_noval = $op(f, preptup_noval..., ba, x, t, contexts...)
+                res2_out2_noval = $op(f, preptup_noval..., ba, x, t, contexts...)
                 res1_out1_val, res2_out1_val = $val_and_op(
-                    f, preptup_noval..., ba, x, tang, contexts...
+                    f, preptup_noval..., ba, x, t, contexts...
                 )
                 res1_out2_val, res2_out2_val = $val_and_op(
-                    f, preptup_noval..., ba, x, tang, contexts...
+                    f, preptup_noval..., ba, x, t, contexts...
                 )
                 let (≈)(x, y) = isapprox(x, y; atol, rtol)
                     @test isempty(preptup_noval) || only(preptup_noval) isa $P
@@ -902,8 +781,8 @@ for op in ALL_OPS
                     end
                 end
             end
-            @test_throws PME $val_and_op(nothing, prepstrict, ba, x, tang, contexts...)
-            @test_throws PME $op(nothing, prepstrict, ba, x, tang, contexts...)
+            @test_throws PME $val_and_op(nothing, prepstrict, ba, x, t, contexts...)
+            @test_throws PME $op(nothing, prepstrict, ba, x, t, contexts...)
             scenario_intact && @test new_scen == scen
             return nothing
         end
@@ -916,35 +795,25 @@ for op in ALL_OPS
             rtol::Real,
             scenario_intact::Bool,
             sparsity::Bool,
+            reprepare::Bool,
         )
-            (; f, x, y, tang, res1, res2, contexts, smaller) = new_scen = deepcopy(scen)
-            xrand, tangrand = myrandom(x), myrandom(tang)
-            rewrap = Rewrap(contexts...)
-            contextsrand = rewrap(map(myrandom ∘ unwrap, contexts)...)
+            (; f, x, y, t, res1, res2, contexts, prep_args) = new_scen = deepcopy(scen)
             local prepstrict
             preptup_cands_val, preptup_cands_noval = map(1:2) do _
-                new_smaller = if isnothing(smaller) || adapt_batchsize(ba, smaller) != ba
-                        deepcopy(scen)
-                    else
-                        deepcopy(smaller)
-                    end
-                prep = $prep_op(f, ba, xrand, tangrand, contextsrand...)
-                prepstrict = $prep_op!(
+                prep = $prep_op(f, ba, prep_args.x, prep_args.t, prep_args.contexts...)
+                prepstrict = $prep_op(
                     f,
-                    $prep_op(
-                        new_smaller.f,
-                        ba,
-                        new_smaller.x,
-                        new_smaller.tang,
-                        new_smaller.contexts...;
-                        strict=Val(true),
-                    ),
                     ba,
-                    xrand,
-                    tangrand,
-                    contextsrand...,
+                    prep_args.x,
+                    prep_args.t,
+                    prep_args.contexts...;
+                    strict=Val(true),
                 )
-                prep_same = $prep_op_same(f, ba, x, tangrand, contexts...)
+                prep_same = $prep_op_same(f, ba, x, map(zero, t), contexts...)
+                if reprepare && has_size(x) && has_size(y) && (size(x) != size(prep_args.x))
+                    prep = $prep_op!(f, prep, ba, x, t, contexts...)
+                    prepstrict = $prep_op!(f, prepstrict, ba, x, t, contexts...)
+                end
                 [(), (prep,), (prepstrict,), (prep_same,)]
             end
             for (preptup_val, preptup_noval) in zip(preptup_cands_val, preptup_cands_noval)
@@ -953,30 +822,16 @@ for op in ALL_OPS
                 res1_in1_val, res2_in1_val = mysimilar(res1), mysimilar(res2)
                 res1_in2_val, res2_in2_val = mysimilar(res1), mysimilar(res2)
                 res2_out1_noval = $op!(
-                    f, res2_in1_noval, preptup_noval..., ba, x, tang, contexts...
+                    f, res2_in1_noval, preptup_noval..., ba, x, t, contexts...
                 )
                 res2_out2_noval = $op!(
-                    f, res2_in2_noval, preptup_noval..., ba, x, tang, contexts...
+                    f, res2_in2_noval, preptup_noval..., ba, x, t, contexts...
                 )
                 res1_out1_val, res2_out1_val = $val_and_op!(
-                    f,
-                    res1_in1_val,
-                    res2_in1_val,
-                    preptup_noval...,
-                    ba,
-                    x,
-                    tang,
-                    contexts...,
+                    f, res1_in1_val, res2_in1_val, preptup_noval..., ba, x, t, contexts...
                 )
                 res1_out2_val, res2_out2_val = $val_and_op!(
-                    f,
-                    res1_in2_val,
-                    res2_in2_val,
-                    preptup_noval...,
-                    ba,
-                    x,
-                    tang,
-                    contexts...,
+                    f, res1_in2_val, res2_in2_val, preptup_noval..., ba, x, t, contexts...
                 )
                 let (≈)(x, y) = isapprox(x, y; atol, rtol)
                     @test isempty(preptup_noval) || only(preptup_noval) isa $P
@@ -984,30 +839,23 @@ for op in ALL_OPS
                     @test res1_in2_val === res1_out2_val
                     @test res1_out1_val ≈ scen.res1
                     @test res1_out2_val ≈ scen.res1
+                    @test res2_in1_noval === res2_out1_noval
+                    @test res2_in2_noval === res2_out2_noval
+                    @test res2_in1_val === res2_out1_val
+                    @test res2_in2_val === res2_out2_val
                     for b in eachindex(scen.res2)
-                        @test res2_in1_noval[b] === res2_out1_noval[b]
-                        @test res2_in2_noval[b] === res2_out2_noval[b]
                         @test res2_out1_noval[b] ≈ scen.res2[b]
                         @test res2_out2_noval[b] ≈ scen.res2[b]
-                        @test res2_in1_val[b] === res2_out1_val[b]
-                        @test res2_in2_val[b] === res2_out2_val[b]
                         @test res2_out1_val[b] ≈ scen.res2[b]
                         @test res2_out2_val[b] ≈ scen.res2[b]
                     end
                 end
             end
             @test_throws PME $op!(
-                nothing, mysimilar(res2), prepstrict, ba, x, tang, contexts...
+                nothing, mysimilar(res2), prepstrict, ba, x, t, contexts...
             )
             @test_throws PME $val_and_op!(
-                nothing,
-                mysimilar(res1),
-                mysimilar(res2),
-                prepstrict,
-                ba,
-                x,
-                tang,
-                contexts...,
+                nothing, mysimilar(res1), mysimilar(res2), prepstrict, ba, x, t, contexts...
             )
             scenario_intact && @test new_scen == scen
             return nothing